add `conftest.py` for scoring tests
